pyproject.toml
[project]
name = "security-rag-assistant"
version = "0.1.0"
description = "Vendor-neutral Security RAG assistant for security log analysis"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "python-dateutil>=2.9",
  "faiss-cpu>=1.8.0",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "typer>=0.12",
]

[project.optional-dependencies]
llm = [
  "openai>=1.47"
]

[project.scripts]
security-rag = "security_rag.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
